Refactor microbenchmark tools: move format_throughput to a new formatting module and update imports
