+++ tests/microbenchmark/tools/summarize_regressions.py
@@ -28,7 +28,7 @@ from pathlib import Path
 import yaml
 
 sys.path.insert(0, str(Path(__file__).resolve().parent))
-from analyze_results import format_throughput  # noqa: E402
+from formatting import format_throughput  # noqa: E402
 
 # nanobench renders unset floating-point fields as bare `-nan`/`nan`, which
 # Python's json parser rejects (it only accepts the capitalized `NaN`).
